review: add a bunch of TODO
The new version looks better!

I'm still worried about the scaling: how it will be able to do all these
operations in userspace per packet for many clients. But if this is the
only way that is possible here, this perf can be improved later (that
will increase the complexity, but this is justified and very likely
required).

Signed-off-by: Matthieu Baerts <matttbe@kernel.org>